Refactoring unit tests to reduce duplication
I extracted the functionality for setting up the Apache LocalTestServer
into a base class that can be extended.
