+++ src/test/java/edu/ksu/canvas/oauth/OauthTokenRefresherUTest.java
@@ -1,11 +1,6 @@
 package edu.ksu.canvas.oauth;
 
-import edu.ksu.canvas.util.JsonTestUtil;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.localserver.LocalTestServer;
-import org.apache.http.protocol.HttpRequestHandler;
-import org.junit.After;
-import org.junit.Before;
+import edu.ksu.canvas.LocalServerTestBase;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
@@ -13,39 +8,14 @@ import org.junit.runners.JUnit4;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 
+import java.util.Collections;
+
 @RunWith(JUnit4.class)
-public class OauthTokenRefresherUTest {
-
-    private LocalTestServer server = new LocalTestServer(null, null);
-    private String baseUrl;
-
-    // Take a JSON file and an HTTP status code and construct a mock response
-    private HttpRequestHandler makeRequestHandlerFromJson(String jsonFileName, int statusCode) {
-        String jsonContent = JsonTestUtil.loadJson(jsonFileName, OauthTokenRefresherUTest.class);
-        return (request, response, context) -> {
-            response.setStatusCode(statusCode);
-            response.setEntity(new StringEntity(jsonContent));
-        };
-    }
-
-    private void registerOauthHandler(HttpRequestHandler handler) {
-        server.register("/login/oauth2/*", handler);
-    }
-
-    @Before
-    public void setUp() throws Exception {
-        server.start();
-        baseUrl = "http:/" + server.getServiceAddress();
-    }
-
-    @After
-    public void tearDown() throws Exception {
-        server.stop();
-    }
+public class OauthTokenRefresherUTest extends LocalServerTestBase {
 
     @Test
     public void successfulRefresh() throws Exception {
-        registerOauthHandler(makeRequestHandlerFromJson("/SampleJson/oauth/SuccessfulAccessTokenResponse.json", 200));
+        registerUrlResponse("/login/oauth2/*", "/SampleJson/oauth/SuccessfulAccessTokenResponse.json", 200, Collections.emptyMap());
         OauthTokenRefresher refresher = new OauthTokenRefresher("", "", baseUrl);
 
         TokenRefreshResponse response = refresher.getNewToken("abc");
@@ -55,7 +25,7 @@ public class OauthTokenRefresherUTest {
 
     @Test
     public void invalidTokenReturnsNull() throws Exception {
-        registerOauthHandler(makeRequestHandlerFromJson("/SampleJson/oauth/InvalidRefreshTokenResponse.json", 400));
+        registerUrlResponse("/login/oauth2/*", "/SampleJson/oauth/InvalidRefreshTokenResponse.json", 400, Collections.emptyMap());
         OauthTokenRefresher refresher = new OauthTokenRefresher("", "", baseUrl);
 
         TokenRefreshResponse response = refresher.getNewToken("");
@@ -65,7 +35,7 @@ public class OauthTokenRefresherUTest {
 
     @Test
     public void invalidClientIdReturnsNull() throws Exception {
-        registerOauthHandler(makeRequestHandlerFromJson("/SampleJson/oauth/InvalidClientIdResponse.json", 401));
+        registerUrlResponse("/login/oauth2/*", "/SampleJson/oauth/InvalidClientIdResponse.json", 401, Collections.emptyMap());
         OauthTokenRefresher refresher = new OauthTokenRefresher("", "", baseUrl);
 
         TokenRefreshResponse response = refresher.getNewToken("");
